augmented cli interface + packaging
