--- db_create.py
@@ -1,15 +1,0 @@
-from migrate.versioning import api
-from config import SQLALCHEMY_DATABASE_URI
-from config import SQLALCHEMY_MIGRATE_REPO
-from app import db
-import os.path
-def create_db():
-	db.create_all()
-	if not os.path.exists(SQLALCHEMY_MIGRATE_REPO):
-	    api.create(SQLALCHEMY_MIGRATE_REPO, 'database repository')
-	    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
-	else:
-	    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO, api.version(SQLALCHEMY_MIGRATE_REPO))
-
-if __name__ == '__main__':
-	create_db()
