--- make_reqs.sh
@@ -1,2 +1,0 @@
-pip freeze > requirements.txt
-echo "psycopg2==2.4.4" >> requirements.txt
