--- redis-start.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-if [ ! -d redis-stable/src ]; then
-    curl -O http://download.redis.io/redis-stable.tar.gz
-    tar xvzf redis-stable.tar.gz
-    rm redis-stable.tar.gz
-fi
-cd redis-stable
-make
-src/redis-server
-
