+++ requirements.txt
@@ -14,6 +14,7 @@ Flask-Mail==0.9.1
 Flask-SQLAlchemy==2.1
 Flask-WhooshAlchemy==0.56
 Flask-WTF==0.14.2
+-e git+https://github.com/james-willis/gearfinder.git@234490bbc1eb7f59b24ea9f63cc46b1f2e8e6b3d#egg=gearfinder
 gunicorn==19.6.0
 itsdangerous==0.24
 Jinja2==2.9.4
